+++ NEWS.md
@@ -1,5 +1,7 @@
 # forestTIME (development version)
 
+- The internal function `fia_eval_info()` was added.
+- `fia_assign_strata()` was added to match plots in each year of the annualized data to an EVALID, estimation unit, and stratum along with associated values necessary for population level estimation with variance calculations.
 - `fia_tidy()` now adds several columns from the POP_* tables relating to EVALIDs.  Each plot in each INVYR is matched to the EVALID of each type that has the latest end year.  Then, for EVALIDs ending in 01 and 02 ("EXPCURR" and "EXPVOL", respectively), two columns are created for each of EVALID, ESTN_UNIT, STRATUMCD, P1POINTCNT, P1PNTCNT_EU, and AREA_USED suffixed with _EXPVOL or _EXPCURR.  There are also two logical columns EVAL_TYPE_EXPCURR and EVAL_TYPE_EXPVOL.  These get filled down and switch at the next inventory (rather than at the midpoint like everything else) (#192). 
 - `fia_tidy()` no longer filters to base intensity plots and reverts to just filtering to `INVYR≥2000`.
 - Changeed the way composit IDs `plot_ID` and `tree_ID` are constructed by changing the order of STATECD and UNITCD so that now it is UNITCD_STATECD_COUNTYCD_PLOT ([#189](https://github.com/Evans-Ecology-Lab/forestTIME/issues/189))
@@ -9,7 +11,6 @@
   - `interpolate_data()` (and therefore `fia_annualize()`) now linearly interpolates `CARBON_AG` and `DRYBIO_AG`
   - `fia_estimate()` now overwrites `CARBON_AG` and `DRYBIO_AG` columns in the input unless the result of carbon estimation is `NA` (as is the case with woodland species).
 - The calculation of the `EXPNS` column no longer occurs as part of `interpolate_data()`.
-- Added a utility function, `fia_calc_expns()` to calculate and add an `EXPNS` column.
 
 # forestTIME 2.1.0